refactor: move Iron utility to IdpUtilities library
- Moved Iron utility from IdentityProvider to IdpUtilities library.
- Updated AuthorizationCallbackController, AuthorizationController, and TokenController to use IdpUtilities.
- Added unit tests for Iron utility in IdpUtilities.Test.
