+++ IdentityProvider/Controllers/AuthorizationCallbackController.cs
@@ -1,4 +1,5 @@
 using IdentityProvider.Models;
+using IdpUtilities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Web;
